refactor(test): Move utils to `test/common/`
